Working Settings Menu
Sorry for the metadata----however, I successfully created a persistent settings menu. The meta data is mainly due to the fact that I had to build the program to test it, and I'm guessing Unity has to do a lot behind the scenes to make that happen.

My main contributions include:

Settings.cs
PanelToggle.cs
ChangeScene.cs

I also did a whole lot of design in the Unity editor to make it all visually appealing, which unfortunately also contributed to the metadata spam.