Started Cow Code and More
The Cow Code was started, with the sheep code being the base. The function that looks for grass has removed for a more universal function that looks for the food types the animal eats. It compares the distance between all of the possible foods and goes to the closest one. A new cow sprite has also been added.