Fixing Chests P.1
Ok so I got it so that chests do not immediatley delete all items. I figured that Serialized Field instiantites using blanks. I do not know if this is common behavior or not but something to note. However, this fix is rendered null as still opening an inventory turns everything null